perf calculations moved into uint64_t
git-svn-id: http://svn.fastmq.org/svn/zeromq/branches/refactoring-for-0.3@350 7451a71c-b6c8-49cc-8254-141fc5361860

+++ perf/helpers/time.hpp
@@ -53,8 +53,8 @@ namespace perf
     {
         struct timespec ts;
         int rc = clock_gettime (CLOCK_MONOTONIC, &ts);
-        assert (rc ==0);
-        return ts.tv_sec * 1000000000 + ts.tv_nsec;
+        assert (rc == 0);
+        return ts.tv_sec * (uint64_t) 1000000000 + ts.tv_nsec;
     }
 
 #if (defined (__GNUC__) && (defined (__i386__) || defined (__x86_64__)))
